Created generic WeakNotifyPropertyChangedProxy
