+++ src/Controls/src/Core/Internals/WeakEventProxy.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Specialized;
+using System.ComponentModel;
 
 // NOTE: warning disabled for netstandard projects
 #pragma warning disable 0436
@@ -99,6 +100,55 @@ namespace Microsoft.Maui.Controls
 			{
 				s.CollectionChanged -= OnCollectionChanged;
 			}
+			base.Unsubscribe();
+		}
+	}
+
+	/// <summary>
+	/// A "proxy" class for subscribing INotifyPropertyChanged via WeakReference.
+	/// General usage is to store this in a member variable and call Subscribe()/Unsubscribe() appropriately.
+	/// Your class should have a finalizer that calls Unsubscribe() to prevent WeakNotifyCollectionChangedProxy objects from leaking.
+	/// </summary>
+	class WeakNotifyPropertyChangedProxy : WeakEventProxy<INotifyPropertyChanged, PropertyChangedEventHandler>
+	{
+		public WeakNotifyPropertyChangedProxy() { }
+
+		public WeakNotifyPropertyChangedProxy(INotifyPropertyChanged source, PropertyChangedEventHandler handler)
+		{
+			Subscribe(source, handler);
+		}
+
+		void OnPropertyChanged(object? sender, PropertyChangedEventArgs e)
+		{
+			if (TryGetHandler(out var handler))
+			{
+				handler(sender, e);
+			}
+			else
+			{
+				Unsubscribe();
+			}
+		}
+
+		public override void Subscribe(INotifyPropertyChanged source, PropertyChangedEventHandler handler)
+		{
+			if (TryGetSource(out var s))
+			{
+				s.PropertyChanged -= OnPropertyChanged;
+			}
+
+			source.PropertyChanged += OnPropertyChanged;
+
+			base.Subscribe(source, handler);
+		}
+
+		public override void Unsubscribe()
+		{
+			if (TryGetSource(out var s))
+			{
+				s.PropertyChanged -= OnPropertyChanged;
+			}
+
 			base.Unsubscribe();
 		}
 	}
